A rule based chatbot
The main motive behind this project was that sometimes students face issues in finding details about the college so to make the process easy we thought of making a chatbot which will easily solve the queries.
This chatbot helps us to know about our university, clubs present in the university and details of the hostels.
So, in this chatbot it first starts the conversation with a pleasant welcome and after greeting we can 
also start by asking our questions to the chatbot. It will help us by providing answers to our 
questions and if the question doesn't match with its database, it will tell us that the sentence was 
not understandable to it.